+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [1.1.0] - 2020-01-15
+
+* Responsive() changed to ResponsiveRow() 
+* Col() changed to ResponsiveColumn() 
+* readme && images updated
+
 ## [1.0.0] - 2020-01-06
 
 * stable release of responsive_ui